cambios uint

+++ src/engine/ECS/Entity/Entity.h
@@ -4,14 +4,14 @@ class EntityManager;
 class Entity
 {
 private:
-    int nextId = 0; ///< Entero que se utiliza para generar la siguiente id unica para una entidad
+    uint16_t nextId = 0; ///< Entero que se utiliza para generar la siguiente id unica para una entidad
     EntityManager *parent;
 
 public:
     /** * @brief Alias para el identificador de una entidad.
      * Se usa un entero simple para identificar cada objeto del juego.
      */
-    using EntityID = int;
+    using EntityID = uint16_t;
 
     Entity(EntityManager* parent) : parent(parent) {}
 
@@ -19,6 +19,6 @@ public:
     EntityManager* getParent(){return parent;}
 
 
-    int getNextId(){return nextId;}
-    void setId(int newId) { nextId = newId; }
+    uint16_t getNextId(){return nextId;}
+    void setId(uint16_t newId) { nextId = newId; }
 };
